requirements.txt
pillow~=10.3.0
setuptools~=60.2.0